feat: worked on landing page
